Move main pipeline object imports to subpackage level
